refactor: cart component 구조 변경
