+++ src/refactoring/components/cart/cart/CouponApply.tsx
@@ -1,6 +1,6 @@
-import { Coupon } from '../../../types';
-import { formatDiscountValue } from '../../services/coupon';
-import { useCouponStore } from '../../stores/useCouponStore';
+import { Coupon } from '../../../../types';
+import { formatDiscountValue } from '../../../services/coupon';
+import { useCouponStore } from '../../../stores/useCouponStore';
 
 interface Props {
   applyCoupon: (coupon: Coupon) => void;
